Add the notification job to let nearby users know when a vehicle is stolen.
This job is written as a Cloud Code job. It should be ran every minute or so.

This will notify nearby users and add them to the chatroom for the stolen vehicle, if they are not already there. Note that I've only tested this to a limited extent: we will need to fake a nearby user for this to work. @darasok let me know when you'd like to give this a test to see if it behaves as it should.